+++ ESPlus/Repositories/InMemoryRepository.cs
@@ -63,7 +63,7 @@ namespace ESPlus.Repositories
             return Task.WhenAll();
         }
 
-        public Task<TAggregate> GetByIdAsync<TAggregate>(string id, int version = int.MaxValue) where TAggregate : IAggregate
+        public Task<TAggregate> GetByIdAsync<TAggregate>(string id, long version = int.MaxValue) where TAggregate : IAggregate
         {
             var instance = (TAggregate)Activator.CreateInstance(typeof(TAggregate), id);
             var aggregate = (IAggregate)instance;
